refactor: split snapshot client and add schema client
