+++ packages/cli/src/lib/services/seed/collections/collection.ts
@@ -4,14 +4,13 @@ import pino from 'pino';
 import { getChildLogger } from '../../../helpers';
 import { SeedMeta, SeedData } from '../interfaces';
 import { SeedDataMapper } from './data-mapper';
-import { SeedIdMapperClient, SeedIdMapperClientFactory } from '../global';
+import { SchemaClient, SeedIdMapperClient, SeedIdMapperClientFactory } from '../global';
 import { SeedDataClient } from './data-client';
 import {
   DirectusId,
   WithSyncId,
   unwrapDirectusRequestError,
 } from '../../collections';
-import { SnapshotClient } from '../../snapshot';
 import { SeedDataDiffer } from './data-differ';
 import { DirectusUnknownType } from '../../interfaces';
 
@@ -28,7 +27,7 @@ export class SeedCollection {
     protected readonly idMapperFactory: SeedIdMapperClientFactory,
     protected readonly dataClient: SeedDataClient,
     protected readonly dataDiffer: SeedDataDiffer,
-    protected readonly snapshotClient: SnapshotClient,
+    protected readonly schemaClient: SchemaClient,
   ) {
     this.logger = getChildLogger(baseLogger, collection);
     this.idMapper = this.idMapperFactory.forCollection(collection);
@@ -226,7 +225,7 @@ export class SeedCollection {
    * Get the primary field name
    */
   protected async getPrimaryFieldName(): Promise<string> {
-    return (await this.snapshotClient.getPrimaryField(this.collection)).name;
+    return (await this.schemaClient.getPrimaryField(this.collection)).name;
   }
 
   /**
