+++ packages/cli/src/lib/services/seed/global/index.ts
@@ -1,3 +1,4 @@
 export * from './id-mapper-client-factory';
 export * from './seed-loader';
 export * from './seed-client';
+export * from './schema-client';
