+++ packages/cli/src/lib/services/snapshot/index.ts
@@ -1,15 +1,2 @@
-export type { Snapshot } from './interfaces';
-export {
-  Accountability,
-  Collapse,
-  Group,
-  Display,
-  Template,
-  Width,
-  DataType,
-  ForeignKeyColumn,
-  Type,
-  OneDeselectAction,
-  On,
-} from './interfaces';
+export * from './interfaces';
 export * from './snapshot-client';
