+++ packages/cli/src/lib/services/snapshot/snapshot-client.ts
@@ -9,10 +9,9 @@ import {
   Relation,
   SchemaDiffOutput,
   Snapshot,
-  Type,
 } from './interfaces';
 import { mkdirpSync, readJsonSync, removeSync, writeJsonSync } from 'fs-extra';
-import { DIRECTUS_COLLECTIONS_PREFIX, LOGGER } from '../../constants';
+import { LOGGER } from '../../constants';
 import pino from 'pino';
 import { getChildLogger, loadJsonFilesRecursively } from '../../helpers';
 import { ConfigService, SnapshotHooks } from '../config';
@@ -128,7 +127,7 @@ export class SnapshotClient {
    * Get the snapshot from the Directus instance.
    */
   @Cacheable()
-  protected async getSnapshot(): Promise<Snapshot> {
+  async getSnapshot(): Promise<Snapshot> {
     const directus = await this.migrationClient.get();
     return await directus.request<Snapshot>(schemaSnapshot()); // Get better types
   }
@@ -278,74 +277,5 @@ export class SnapshotClient {
       const filePath = path.join(this.dumpPath, SNAPSHOT_JSON);
       return readJsonSync(filePath, 'utf-8') as Snapshot;
     }
-  }
-
-  /**
-   * Returns the list of fields of a model.
-   */
-  @Cacheable()
-  async getFields(model: string): Promise<Field[]> {
-    const snapshot = await this.getSnapshot();
-    return snapshot.fields.filter((f) => f.collection === model);
-  }
-
-  /**
-   * Returns the list of fields of type "many-to-one" of a model.
-   */
-  @Cacheable()
-  async getRelationFields(model: string): Promise<string[]> {
-    const snapshot = await this.getSnapshot();
-    return snapshot.relations
-      .filter((r) => r.collection === model)
-      .map((r) => r.field);
-  }
-
-  /**
-   * Returns the target model of a many-to-one relation.
-   * Throw an error if the relation is not many-to-one
-   * or the model does not exist or the field does not exist.
-   */
-  @Cacheable()
-  async getTargetModel(model: string, field: string): Promise<string> {
-    const snapshot = await this.getSnapshot();
-    const relation = snapshot.relations.find(
-      (r) => r.collection === model && r.field === field,
-    );
-    if (!relation) {
-      throw new Error(
-        `Relation ${model}.${field} does not exist in the snapshot`,
-      );
-    }
-    return relation.related_collection;
-  }
-
-  /**
-   * Returns the primary field of a model.
-   */
-  @Cacheable()
-  async getPrimaryField(model: string): Promise<{ name: string; type: Type }> {
-    if (model.startsWith(DIRECTUS_COLLECTIONS_PREFIX)) {
-      return this.getDirectusCollectionPrimaryField(model);
-    }
-    const snapshot = await this.getSnapshot();
-    const fields = snapshot.fields.filter((c) => c.collection === model);
-    const primaryField = fields.find((f) => f.schema?.is_primary_key);
-    if (!primaryField) {
-      throw new Error(`Primary field not found in ${model}`);
-    }
-    return { name: primaryField.field, type: primaryField.type };
-  }
-
-  /**
-   * Returns the primary field type of a directus model.
-   */
-  protected getDirectusCollectionPrimaryField(model: string): {
-    name: string;
-    type: Type;
-  } {
-    if (['directus_permissions', 'directus_presets'].includes(model)) {
-      return { name: 'id', type: Type.Integer };
-    }
-    return { name: 'id', type: Type.UUID };
   }
 }
